Use file storage to store volumes when no non-OS disks is selected
When additional non-OS disks are available, it's possible to select one
or more of them to use them as storage for volumes. If none of them is
selected, then one is automatically chosen.

This "automatically chosen" feature can be an issue (the disk can be
used for something else), and people might still want to use some file
storage (which is what happens when no non-OS disk is available). So
change the behavior to use file storage if no disk is selected.

Part of https://bugzilla.novell.com/show_bug.cgi?id=774317