Add compute hosts to their availability zone
There's a new crowbar-nova-set-availability-zone utility that will be
run on the controller node to configure the availability zone of compute
nodes. It will only do so when the compute nodes are known to the
controller nodes, which might happen after the first chef-client run;
but 15 minutes later, everything will go fine.