Run crowbar-nova-set-availability-zone command in a delayed notification
Until now, we were running the crowbar-nova-set-availability-zone calls
as part of the chef run, which resulted in a loss of nova-related
notification if the calls ever failed. This could result in services not
being restarted after a config change.

So instead, we use the same trick as in the neutron cookbook and we run
the commands in a delayed notification, so that it runs at the end.

This makes us lose the synchronization, but I don't think it's strictly
required here.

(Failures might still impact notifications of cookbooks coming later,
but that's less likely than notifications for the nova services since a
failure here would generally be caused by a change in the nova config,
not by a change in the config from later cookbooks)