Use crowbar built-in reboot handler for xen kernel
When applying the Xen role, the compute node needs to reboot into the
Xen kernel flavor. Crowbars service object has a built-in reboot handler
which waits until the reboot is completed. This has the nice side effect
that the proposal commit is finished when the reboot is done. That's
currently not the case.

https://bugzilla.novell.com/show_bug.cgi?id=878012