Generate certs on all nodes for HA
With HA, HTTPS and automatic generation of self-signed certificates, all
controller nodes except for one are missing the cert files, and the api
and novncproxy services fail to start.

This is a workaround to generate the certs on all nodes when HA is
enabled.
(cherry picked from commit 7cc8e23b67695ea3ce05088bebec30095e6178db)