Move calls to crowbar-nova-set-availability-zone to compute nodes
Instead of doing N calls (where N is the number of compute nodes) on the
controller, which doesn't scale (it can be slow and make chef-client
take a long time to complete), we just do one call on each compute node.

Except for Hyper-V compute nodes, that are still handled by the
controller due to the fact that they don't have the small utility.