Add option to enforce use of config drive (bsc#943166, bsc#941531)
This is useful in environments where the dhcp-agent is
disabled, which means there is no neutron/nova metadata proxy
available. Also inject the flat network in that case, so
that cloud-init can pick up network configuration in case
DHCP is turned off.

Backport of https://github.com/crowbar/crowbar-openstack/pull/17

Co-Authored-By: Thomas Bechtold <tbechtold@suse.com>